RF: move+RF code for navigating dandi archive url from download into dandiarchive
So that it could be used also for ls or any other code which needs to
go through the thing (dandiset, folder, assets()) in the archive.
ATM we need to flip between Girder and DANDI API, so hat helper abstracts
that away a bit more on top of having GirderCli and DandiAPIClient
